requirements.txt
django
PyYAML
bs4